refactored pharmacy operations from main to setup package, moved DataProvider to setup package, implemented cartsupplier and tests, customers have carts, work on ICashier
