+++ src/main/java/inventory/Cart.java
@@ -1,11 +1,25 @@
 package inventory;
 
 import java.util.LinkedHashMap;
+import product.Product;
 
 public class Cart extends Inventory {
 
     public Cart() {
         this.products = new LinkedHashMap<>();
+    }
+
+    public static Cart cartSupplier(ProductInventory productInventory, int numItems) {
+        Cart cart = new Cart();
+        for(int i = 0; i < numItems; i++){
+            Product randomProduct = productInventory.getRandomProduct();
+            cart.addProduct(randomProduct, 1);
+        }
+        return cart;
+    }
+
+    public int getNumberOfItemsInCart(){
+        return this.products.values().stream().mapToInt(Integer::intValue).sum();
     }
 
     @Override
