+++ src/main/java/person/ICashier.java
@@ -1,0 +1,9 @@
+package person;
+
+import genericLinkedList.CustomerLine;
+
+public interface ICashier {
+    default AbstractCustomer getCustomerFromCustomerLine(CustomerLine customerLine){
+        return customerLine.getNextCustomer();
+    }
+}
