+++ src/test/java/PharmacistTest.java
@@ -1,4 +1,4 @@
-import misc.DataProvider;
+import setup.DataProvider;
 import person.Pharmacist;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
